LayoutTests/imported/w3c/web-platform-tests/uievents/mouse/layer-coords-transform-expected.txt


FAIL layer-coords-transform assert_equals: layerX expected 200 but got 210
PASS layer-coords-transform 1
